+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Other:
   - Decreased build size by not importing `pouchdb-utils`
+  - Improve build size and performance by replacing [deep-equal](https://www.npmjs.com/package/deep-equal) with [fast-deep-equal](https://github.com/epoberezkin/fast-deep-equal#readme)
 
 ### 10.4.0 (11 November 2021)